Change trigger for splitscreen toggle
Fixes #1 . Splitscreen was previously toggled by the FOV, which *worked*, but was janky. And it seems that due to Cinemachine updates, this no longer works. Splitscreen is now toggled when the players are far enough apart, which should have a similar effect. There are probably some edge case issues that haven't been tested, though.